+++ react-app/src/store/game.js
@@ -1,25 +1,10 @@
 const GET_GAMES = 'games/GET_GAMES';
-const GET_COURSES = 'courses/GET_COURSES'
-const GET_COURSE_INFO = 'courses/GET_COURSE_INFO'
+
 
 const loadGames = (games) => {
     return {
         type: GET_GAMES,
         games,
-    }
-}
-
-const loadCourses = (courses) => {
-    return {
-        type: GET_COURSES,
-        courses,
-    }
-}
-
-const loadCourseInfo = (courseInfo) => {
-    return {
-        type: GET_COURSE_INFO,
-        courseInfo,
     }
 }
 
@@ -31,26 +16,6 @@ export const getGames = () => async (dispatch) => {
         const games = await response.json()
         console.log(games)
         await dispatch(loadGames(games))
-        return response
-    }
-}
-
-export const getCoursesForGame = (gameId) => async (dispatch) => {
-    const response = await fetch(`/api/game/${gameId}`)
-
-    if (response.ok) {
-        const courses = await response.json()
-        await dispatch(loadCourses(courses))
-        return response
-    }
-}
-
-export const getCourseInfo = (courseId) => async (dispatch) => {
-    const response = await fetch(`/api/course/${courseId}`)
-
-    if (response.ok) {
-        const courseInfo = await response.json()
-        await dispatch(loadCourseInfo(courseInfo))
         return response
     }
 }
@@ -70,20 +35,6 @@ export default function games(state = initialState, action) {
             return newState;
         }
 
-        case GET_COURSES: {
-            const allCourses = {};
-            action.courses.courses.forEach(course => {
-                allCourses[course.id] = course;
-            });
-            newState = {...allCourses}
-            return newState;
-        }
-
-        case GET_COURSE_INFO: {
-            newState = Object.assign({}, state);
-            newState.courseInfo = action.courseInfo;
-            return newState;
-        }
 
         default:
             return state
